feat: add queries for quiz(#109)
