+++ strawberry/src/pages/quizLanding/hooks/index.tsx
@@ -1,12 +1,10 @@
 import { useTimeConversion } from "./useTimeConversion";
-import { useQuizLandingQuery } from "./useQuizLandingQuery";
 import { useQuizLandingDispatch } from "./useQuizLandingDispatch";
 import { useQuizLandingState } from "./useQuizLandingState";
 import { useQuizLandingData } from "./useQuizLandingData";
 
 export {
   useTimeConversion,
-  useQuizLandingQuery,
   useQuizLandingDispatch,
   useQuizLandingState,
   useQuizLandingData,
