--- strawberry/src/pages/quizPlay/hooks/queries/useQuizPlayMutation.tsx
@@ -1,34 +1,0 @@
-import { useMutation } from "react-query";
-
-import { useFetch } from "../../../../data/config/useFetch";
-import usePrizeModal from "../usePrizeModal";
-
-export interface UseQuizPlayMutationBodyType {
-  answer: string;
-  subEventId: number;
-}
-
-export function useQuizPlayMutation() {
-  const prizeModal = usePrizeModal();
-
-  const fetchFn = useFetch({
-    url: "firstcome/quiz",
-    method: "POST",
-  });
-
-  const mutation = useMutation({
-    mutationFn: ({ body }: { body: UseQuizPlayMutationBodyType }) =>
-      fetchFn({ body: body }),
-    onSuccess: (data) => {
-      if (data) {
-        prizeModal({
-          isCorrect: data?.isCorrect,
-          isWinner: data?.isWinner,
-          prizeImgUrl: data?.prizeImgUrl,
-        });
-      }
-    },
-  });
-
-  return mutation;
-}
